+++ integrationTest/package.json
@@ -11,10 +11,10 @@
     "grunt-contrib-clean": "^1.0.0",
     "grunt-contrib-jshint": "^1.0.0",
     "install-local": "^0.5.0",
-    "jasmine-core": "^2.4.1",
+    "jasmine-core": "^3.1.0",
     "karma": "^2.0.2",
     "karma-chai": "^0.1.0",
-    "karma-jasmine": "^1.1.1",
+    "karma-jasmine": "^1.1.2",
     "karma-mocha": "^1.3.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "link-parent-bin": "^0.2.0",
