+++ integrationTest/test/mocha-mocha/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": false
+  },
+  "files": [
+    "verify/verify.ts"
+  ]
+}
