+++ packages/stryker-karma-runner/package.json
@@ -37,7 +37,7 @@
   "devDependencies": {
     "jasmine-core": "^2.4.1",
     "karma": "^2.0.0",
-    "karma-jasmine": "^1.1.1",
+    "karma-jasmine": "^1.1.2",
     "karma-phantomjs-launcher": "^1.0.4",
     "stryker-api": "^0.17.0",
     "stryker-jasmine": "^0.8.4"
